# Use cuBLAS for large batches and quants with block size 16