@ewhacc 2025-06-26
I tried this "build = 3773 (3dbc8437)"  on ubergam's  DeepSeek-R1-0528-GGUF  IQ2_K_R4  with  -b 4096 -ub 4096.
Getting no difference on PP speed, compared to "build = 3762 (1843ed22)".

Both are about the same:
prompt eval time =   25328.73 ms /  6889 tokens (    3.68 ms per token,   271.98 tokens per second)

Did I something wrong?

My rig is Epyc Genoa + 6000 ada.

Built with
cmake -B ./build -DGGML_CUDA=ON -DGGML_BLAS=OFF -DGGML_SCHED_MAX_COPIES=1 -DGGML_CUDA_IQK_FORCE_BF16=1