@ubergarm 2025-06-26
@ewhacc 
@ewhacc 

Yeah, the speed boosts specific to IQ2_K_R4 and IQ3_K_R4 quantizations (in the quan you mention) were *already* added in PR557. This PR is doing a similar thing for some *other* quant types like Q2_K etc.

I just did another test for PR557 using this git sha, which is a bit confusing as I'm not actually testing all the new quants added here. But you can see the speed up is pretty good relative to just *before* PR557 was merged as shown below:

![sweep-bench-PR557-revisit](https://github.com/user-attachments/assets/bda70fa0-94a1-4e08-85b6-2850f0fd1815)

<details>

<summary>👈</summary>

```bash
cmake -B ./build -DGGML_CUDA=ON -DGGML_BLAS=OFF -DGGML_SCHED_MAX_COPIES=1 -DGGML_CUDA_IQK_FORCE_BF16=1 -DGGML_CUDA_F16=ON
cmake --build ./build --config Release -j $(nproc)

model=DeepSeek-R1-0528-IQ2_K_R4-00001-of-00005.gguf
./build/bin/llama-sweep-bench \
    --model "$model" \
    --no-mmap \
    --ctx-size 12288 \
    -ctk q8_0 \
    -mla 3 -fa \
    -fmoe \
    -amb 512 \
    -ngl 99 \
    -ot "blk\.(3|4|5|6|7|8|9|10|11|12)\.ffn_.*=CUDA0" \
    -ot "blk\.(13|14|15|16|17|18|19|20|21|22)\.ffn_.*=CUDA1" \
    -ot exps=CPU \
    --warmup-batch \
    --threads 24

llama_model_loader: - type  f32:  361 tensors
llama_model_loader: - type q5_0:   61 tensors
llama_model_loader: - type iq4_ks:  116 tensors
llama_model_loader: - type iq5_ks:  435 tensors
llama_model_loader: - type iq2_k_r4:  116 tensors
llama_model_loader: - type iq3_k_r4:   58 tensors
```

## PR559@3dbc8437 -ub 512 -b 2048
|    PP |     TG |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |
|-------|--------|--------|----------|----------|----------|----------|
|   512 |    128 |      0 |    4.153 |   123.28 |    8.016 |    15.97 |
|   512 |    128 |    512 |    3.844 |   133.18 |    8.126 |    15.75 |
|   512 |    128 |   1024 |    3.932 |   130.22 |    8.246 |    15.52 |
|   512 |    128 |   1536 |    4.104 |   124.74 |    8.179 |    15.65 |
|   512 |    128 |   2048 |    4.185 |   122.35 |    8.188 |    15.63 |
|   512 |    128 |   2560 |    4.265 |   120.04 |    8.452 |    15.14 |
|   512 |    128 |   3072 |    4.576 |   111.89 |    8.376 |    15.28 |
|   512 |    128 |   3584 |    5.258 |    97.37 |    8.491 |    15.07 |
|   512 |    128 |   4096 |    4.538 |   112.83 |    8.456 |    15.14 |
|   512 |    128 |   4608 |    4.625 |   110.69 |    8.483 |    15.09 |
|   512 |    128 |   5120 |    4.717 |   108.55 |    8.609 |    14.87 |
|   512 |    128 |   5632 |    4.796 |   106.76 |    8.704 |    14.71 |
|   512 |    128 |   6144 |    4.950 |   103.42 |    8.862 |    14.44 |
|   512 |    128 |   6656 |    4.939 |   103.66 |    8.781 |    14.58 |
|   512 |    128 |   7168 |    5.195 |    98.55 |    8.722 |    14.68 |
|   512 |    128 |   7680 |    5.062 |   101.14 |    8.778 |    14.58 |
|   512 |    128 |   8192 |    5.199 |    98.49 |    8.962 |    14.28 |

## PR559@3dbc8437 -ub 2048 -b 2048
|    PP |     TG |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |
|-------|--------|--------|----------|----------|----------|----------|
|  2048 |    512 |      0 |    9.450 |   216.73 |   32.442 |    15.78 |
|  2048 |    512 |   2048 |    9.884 |   207.20 |   32.834 |    15.59 |
|  2048 |    512 |   4096 |   10.350 |   197.87 |   33.770 |    15.16 |
|  2048 |    512 |   6144 |   10.742 |   190.65 |   34.733 |    14.74 |
|  2048 |    512 |   8192 |   11.167 |   183.39 |   36.017 |    14.22 |


## PR559@3dbc8437 -ub 4096 -b 4096
|    PP |     TG |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |
|-------|--------|--------|----------|----------|----------|----------|
|  4096 |   1024 |      0 |   12.824 |   319.40 |   65.575 |    15.62 |
|  4096 |   1024 |   4096 |   14.822 |   276.35 |   68.417 |    14.97 |
|  4096 |   1024 |   8192 |   17.282 |   237.01 |   72.403 |    14.14 |

## main@8e5106b2 -ub 512 -b 2048
|    PP |     TG |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |
|-------|--------|--------|----------|----------|----------|----------|
|   512 |    128 |      0 |    4.339 |   117.99 |    8.091 |    15.82 |
|   512 |    128 |    512 |    4.411 |   116.08 |    8.349 |    15.33 |
|   512 |    128 |   1024 |    4.516 |   113.38 |    8.158 |    15.69 |
|   512 |    128 |   1536 |    4.873 |   105.07 |    8.190 |    15.63 |
|   512 |    128 |   2048 |    4.667 |   109.71 |    8.288 |    15.44 |
|   512 |    128 |   2560 |    4.763 |   107.49 |    8.379 |    15.28 |
|   512 |    128 |   3072 |    4.854 |   105.48 |    8.572 |    14.93 |
|   512 |    128 |   3584 |    4.932 |   103.82 |    8.421 |    15.20 |
|   512 |    128 |   4096 |    5.477 |    93.48 |    8.420 |    15.20 |
|   512 |    128 |   4608 |    5.125 |    99.90 |    8.553 |    14.97 |
|   512 |    128 |   5120 |    5.283 |    96.92 |    8.611 |    14.87 |
|   512 |    128 |   5632 |    5.393 |    94.94 |    8.668 |    14.77 |
|   512 |    128 |   6144 |    5.853 |    87.48 |    8.709 |    14.70 |
|   512 |    128 |   6656 |    5.466 |    93.66 |    8.837 |    14.48 |
|   512 |    128 |   7168 |    5.547 |    92.29 |    8.730 |    14.66 |
|   512 |    128 |   7680 |    5.648 |    90.64 |    8.885 |    14.41 |
|   512 |    128 |   8192 |    5.796 |    88.34 |    8.954 |    14.29 |

## main@8e5106b2 -ub 2048 -b 2048
|    PP |     TG |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |
|-------|--------|--------|----------|----------|----------|----------|
|  2048 |    512 |      0 |   11.483 |   178.34 |   32.442 |    15.78 |
|  2048 |    512 |   2048 |   11.937 |   171.56 |   33.131 |    15.45 |
|  2048 |    512 |   4096 |   12.262 |   167.02 |   33.925 |    15.09 |
|  2048 |    512 |   6144 |   12.714 |   161.08 |   34.877 |    14.68 |
|  2048 |    512 |   8192 |   13.044 |   157.01 |   36.298 |    14.11 |


## main@8e5106b2 -ub 4096 -b 4096
|    PP |     TG |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |
|-------|--------|--------|----------|----------|----------|----------|
|  4096 |   1024 |      0 |   14.738 |   277.93 |   65.731 |    15.58 |
|  4096 |   1024 |   4096 |   16.671 |   245.70 |   68.219 |    15.01 |
|  4096 |   1024 |   8192 |   19.206 |   213.26 |   72.408 |    14.14 |


</details>