@ikawrakow 2025-06-27
Here is an example illustrating my previous post. Running LlaMA-3.1-70B quantized with `Q2_K_S` on my paltry RTX-4080 with 16 GB VRAM:

| model            | n_ubatch |     test |    t/s (main)    |    t/s (PR)      |  Speedup |
| ---------------- | -------: | -------: | ---------------: | ---------------: | -------: |
| llama 70B Q2_K   |      512 |   pp4096 |    302.11 ± 0.38 |    328.51 ± 1.02 |  1.087   |
| llama 70B Q2_K   |     1024 |   pp4096 |    397.43 ± 0.36 |    488.37 ± 0.27 |  1.229   |
| llama 70B Q2_K   |     2048 |   pp4096 |    468.44 ± 0.02 |    626.39 ± 0.30 |  1.338   |
| llama 70B Q2_K   |     4096 |   pp4096 |    509.45 ± 0.19 |    722.58 ± 0.40 |  1.418   |

I have uploaded only 30 out of 80 layers to the GPU so I can run with the larger u-batch. If instead I use the default u-batch of 512, I can upload 50 layers to the GPU. With that I get `pp4096 = 372 t/s` on the main branch, so a pretty good speedup with this PR and `u-batch = 4096` with almost double the performance.